requirements.txt
flask
joblib
opencv-python
numpy
scikit-learn
pillow
transformers
torch
